requirements.txt
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cssselect==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
idna==3.4
inflection==0.5.1
jieba3k==0.35.1
joblib==1.2.0
jsonschema==4.17.3
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
Pillow==9.4.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
wikipedia==1.4.0
youtube-transcript-api==0.5.0
